refactor(parser,cli): move findDependents into parser, decoupled from VectorDBInterface
findDependents (the get_dependents MCP tool's engine) lived in cli, the top
of the parser <- core <- cli dependency stack, so packages/review (which
depends on parser only) couldn't reuse it and grew its own weaker,
independently-drifting dependency graph instead.

The VectorDBInterface dependency was an illusion: the cli file used
`import type` for everything from @liendev/core except one call,
vectorDB.scanAll(). SearchResult is a structural superset of CodeChunk, so
making findDependents and its helpers generic over <T extends CodeChunk>
(the same technique #973 applied to addFuzzyMatchChunks/findDependentChunks)
let the whole algorithm move to @liendev/parser unchanged in behavior,
taking Iterable<T> chunks instead of a database handle.

Merged into parser's existing dependency-analyzer.ts (analyzeDependencies's
home) rather than a sibling module, so the two algorithms share their
low-level helpers for real instead of drifting apart in two places again:
createPathNormalizer, groupChunksByFile, buildReExportGraph,
calculateFileComplexities, and calculateComplexityRiskBoost are now single
generic implementations. calculateOverallComplexityMetrics stays split
(analyzeDependencies returns undefined on no data; findDependents needs an
always-present default) -- calculateComplexityMetricsOrDefault wraps the
first rather than duplicating it.

cli's dependency-analyzer.ts is now a thin wrapper: fetch (and cache) chunks
via vectorDB.scanAll(), call parser's findDependents. scanCache stays in cli
(the caller knows its indexVersion); parser has no mutable module state.

No behavioral change -- get_dependents output (dependentCount, riskLevel,
attributionCaveat, and the full dependent-filepath set, file- and
symbol-level) is identical before/after, verified against this repo's index.
